Use the platform as the base interpreter instead of trying to extend it
