+++ packages/interpreter/src/ts/interpreter_core.ts
@@ -76,5 +76,17 @@ export class Interpreter {
     const id = element.getAttribute("data-dioxus-id");
     delete this.local[id];
   }
+
+  getNode(id: number): Node {
+    return this.nodes[id];
+  }
+
+  appendChildren(id: number, many: number) {
+    const root = this.nodes[id];
+    const els = this.stack.splice(this.stack.length - many);
+    for (let k = 0; k < many; k++) {
+      root.appendChild(els[k]);
+    }
+  }
 }
